refactor: ♻️ refactor dataWrokSpace
