+++ apps/builder/src/page/App/components/DataWorkspace/components/ComponentSpaceTree/index.tsx
@@ -1,28 +1,21 @@
-import { cloneDeep } from "lodash"
 import { FC, MouseEvent, useCallback, useMemo } from "react"
 import { useTranslation } from "react-i18next"
 import { useDispatch, useSelector } from "react-redux"
 import { omit } from "@illa-design/react"
 import { PanelBar } from "@/components/PanelBar"
-import { hiddenFields } from "@/page/App/components/DataWorkspace/components/WorkSpaceTree"
 import { WorkSpaceTreeGroup } from "@/page/App/components/DataWorkspace/components/WorkSpaceTreeGroup"
 import { WorkSpaceTreeItem } from "@/page/App/components/DataWorkspace/components/WorkSpaceTreeItem"
-import { changeSelectedDisplayName } from "@/page/App/components/ScaleSquare/utils/changeSelectedDisplayName"
+import { hiddenFields } from "@/page/App/components/DataWorkspace/constant"
 import { getSelectedComponents } from "@/redux/config/configSelector"
 import { configActions } from "@/redux/config/configSlice"
 import { updateCurrentAllComponentsAttachedUsers } from "@/redux/currentApp/collaborators/collaboratorsHandlers"
 import { getComponentAttachUsers } from "@/redux/currentApp/collaborators/collaboratorsSelector"
-import {
-  getComponentDisplayNameMapDepth,
-  getShowWidgetNameParentMap,
-} from "@/redux/currentApp/editor/components/componentsSelector"
 import {
   getGeneralWidgetExecutionResultArray,
   getModalWidgetExecutionResultArray,
 } from "@/redux/currentApp/executionTree/executionSelector"
 import { executionActions } from "@/redux/currentApp/executionTree/executionSlice"
 import { FocusManager } from "@/utils/focusManager"
-import { isMAC } from "@/utils/userAgent"
 
 export const ComponentSpaceTree: FC = () => {
   const { t } = useTranslation()
@@ -36,52 +29,9 @@ export const ComponentSpaceTree: FC = () => {
     getModalWidgetExecutionResultArray,
   )
   const selectedComponents = useSelector(getSelectedComponents)
-  const displayNameMapDepth = useSelector(getComponentDisplayNameMapDepth)
-  const widgetDisplayNameRelationMap = useSelector(getShowWidgetNameParentMap)
-
-  const handleSelectComponentWhenPressMetaKey = useCallback(
-    (selectedKeys: string[]) => {
-      let currentSelectedDisplayName = cloneDeep(selectedComponents)
-      const index = currentSelectedDisplayName.findIndex(
-        (displayName) => displayName === selectedKeys[0],
-      )
-      if (index !== -1) {
-        currentSelectedDisplayName.splice(index, 1)
-      } else {
-        currentSelectedDisplayName.push(...selectedKeys)
-      }
-      changeSelectedDisplayName(
-        currentSelectedDisplayName,
-        widgetDisplayNameRelationMap,
-        selectedKeys[0],
-        displayNameMapDepth,
-      )
-      currentSelectedDisplayName = Array.from(
-        new Set(currentSelectedDisplayName),
-      )
-      dispatch(
-        configActions.updateSelectedComponent(currentSelectedDisplayName),
-      )
-      updateCurrentAllComponentsAttachedUsers(
-        currentSelectedDisplayName,
-        componentsAttachedUsers,
-      )
-    },
-    [
-      componentsAttachedUsers,
-      dispatch,
-      displayNameMapDepth,
-      selectedComponents,
-      widgetDisplayNameRelationMap,
-    ],
-  )
 
   const handleGeneralComponentSelect = useCallback(
     (selectedKeys: string[], e: MouseEvent<HTMLDivElement>) => {
-      // if ((isMAC() && e.metaKey) || (!isMAC() && e.ctrlKey)) {
-      //   handleSelectComponentWhenPressMetaKey(selectedKeys)
-      //   return
-      // }
       dispatch(configActions.updateSelectedComponent(selectedKeys))
       updateCurrentAllComponentsAttachedUsers(
         selectedKeys,
@@ -93,10 +43,6 @@ export const ComponentSpaceTree: FC = () => {
 
   const handleModalComponentSelect = useCallback(
     (selectedKeys: string[], e: MouseEvent<HTMLDivElement>) => {
-      // if ((isMAC() && e.metaKey) || (!isMAC() && e.ctrlKey)) {
-      //   handleSelectComponentWhenPressMetaKey(selectedKeys)
-      //   return
-      // }
       dispatch(
         executionActions.updateModalDisplayReducer({
           displayName: selectedKeys[0],
