--- apps/builder/src/page/App/components/DataWorkspace/components/ComponentSpaceTree/interface.ts
@@ -1,6 +1,0 @@
-import { MouseEvent } from "react"
-
-export interface ComponentSpaceTreeProps {
-  selectedKeys: string[]
-  handleSelect: (selectedKeys: string[], e: MouseEvent<HTMLDivElement>) => void
-}
