--- apps/builder/src/page/App/components/DataWorkspace/components/WorkSpaceTree/interface.ts
@@ -1,9 +1,0 @@
-import { MouseEvent } from "react"
-
-export interface WorkSpaceTreeProps {
-  title: string
-  dataList?: Record<string, any>[]
-  selectedKeys?: string[]
-  handleSelect?: (selectedKeys: string[], e: MouseEvent<HTMLDivElement>) => void
-  onIllaFocus?: () => void
-}
