--- apps/builder/src/page/App/components/DataWorkspace/components/WorkSpaceTree/style.ts
@@ -1,7 +1,0 @@
-import { SerializedStyles, css } from "@emotion/react"
-
-export function applyTreeContainerStyle(): SerializedStyles {
-  return css`
-    font-family: "Fira Code", monospace;
-  `
-}
