+++ apps/builder/src/page/App/components/DataWorkspace/components/WorkSpaceTreeGroup/index.tsx
@@ -1,13 +1,13 @@
 import { FC } from "react"
 import { PanelBar } from "@/components/PanelBar"
-import { applyTreeContainerStyle } from "@/page/App/components/DataWorkspace/components/WorkSpaceTree/style"
-import { WorkSpaceTreeGroupProps } from "@/page/App/components/DataWorkspace/components/WorkSpaceTreeGroup/interface"
+import { WorkSpaceTreeGroupProps } from "./interface"
+import { treeContainerStyle } from "./style"
 
 export const WorkSpaceTreeGroup: FC<WorkSpaceTreeGroupProps> = (props) => {
   const { title, children } = props
   return (
     <PanelBar title={title} size="small">
-      <div css={applyTreeContainerStyle()}>{children}</div>
+      <div css={treeContainerStyle}>{children}</div>
     </PanelBar>
   )
 }
