+++ apps/builder/src/page/App/components/DataWorkspace/components/WorkSpaceTreeGroup/style.ts
@@ -1,0 +1,5 @@
+import { css } from "@emotion/react"
+
+export const treeContainerStyle = css`
+  font-family: "Fira Code", monospace;
+`
