+++ apps/builder/src/page/App/components/DataWorkspace/constant.ts
@@ -1,0 +1,1 @@
+export const hiddenFields = ["displayName", "userId"]
